rename folder according to requirement to lower case
