+++ keyboards/lily58/glow_enc/info.json
