+++ keyboards/lily58/lily58.h
@@ -4,6 +4,6 @@
     #include "rev1.h"
 #elif KEYBOARD_lily58_light
     #include "light.h"
-#elif KEYBOARD_lily58_glowEnc
-    #include "glowEnc.h"
+#elif KEYBOARD_lily58_glow_enc
+    #include "glow_enc.h"
 #endif
